# api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GemWorld B2B Gemstone Marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TransactionStatus =
  (typeof TransactionStatus)[keyof typeof TransactionStatus];

export const TransactionStatus = {
  pending: "pending",
  completed: "completed",
  overdue: "overdue",
} as const;

export type Currency = (typeof Currency)[keyof typeof Currency];

export const Currency = {
  USD: "USD",
  INR: "INR",
  AED: "AED",
} as const;

export interface Transaction {
  id: string;
  buyer_id: string;
  seller_id: string;
  inventory_id: string;
  /** @minimum 0 */
  total_amount: number;
  currency: Currency;
  /** @minimum 0 */
  advance_paid: number;
  /** @minimum 0 */
  credit_amount: number;
  due_date: string;
  status: TransactionStatus;
  created_at: string;
}

export interface CreateTransactionRequest {
  buyer_id: string;
  seller_id: string;
  inventory_id: string;
  /** @minimum 0 */
  total_amount: number;
  currency: Currency;
  /** @minimum 0 */
  advance_paid: number;
  /** @minimum 0 */
  credit_amount: number;
  due_date: string;
}

export interface Payment {
  id: string;
  transaction_id: string;
  /** @minimum 0 */
  amount: number;
  paid_at: string;
}

export interface CreatePaymentRequest {
  transaction_id: string;
  /** @minimum 0 */
  amount: number;
}

export interface CreditSummary {
  buyer_id: string;
  /**
   * Computed credit score (300–850). Higher is better.
   * @minimum 300
   * @maximum 850
   */
  credit_score: number;
  /** Human-readable rating derived from credit score */
  credit_rating: string;
  total_outstanding: number;
  overdue_amount: number;
  completed_transactions: number;
  overdue_transactions: number;
  transactions: Transaction[];
}

/**
 * Confidence level for the estimated price range
 */
export type GemstonePricingConfidence =
  (typeof GemstonePricingConfidence)[keyof typeof GemstonePricingConfidence];

export const GemstonePricingConfidence = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface Gemstone {
  id: string;
  seller_id: string;
  stone_type: string;
  /** @minimum 0 */
  carat: number;
  origin: string;
  treatment: string;
  /** @minimum 0 */
  price: number;
  currency: Currency;
  /** @minimum 0 */
  base_price_usd: number;
  certificate_number: string;
  image_url: string;
  /** GIA colour grade (D–M). Required for diamonds. */
  color?: string | null;
  /** GIA clarity grade (IF, VVS1, VVS2, VS1, VS2, SI1, SI2, I1, I2, I3). Required for diamonds. */
  clarity?: string | null;
  /** Rapaport Benchmark Price (USD/carat). Diamonds only. */
  rap_price_per_carat?: number | null;
  /** Total Rapaport value = rap_price_per_carat × carat. Diamonds only. */
  total_rap_value?: number | null;
  is_featured: boolean;
  boost_expiry_date?: string | null;
  /**
   * Estimated Trade Range — floor (USD). Diamonds: 60% of Rap. Others: rule-based.
   * @minimum 0
   */
  estimated_price_min: number;
  /**
   * Estimated Trade Range — ceiling (USD). Diamonds: 90% of Rap. Others: rule-based.
   * @minimum 0
   */
  estimated_price_max: number;
  /** Confidence level for the estimated price range */
  pricing_confidence: GemstonePricingConfidence;
  /** Disclaimer for the estimated price */
  pricing_disclaimer: string;
  created_at: string;
}

export interface CreateGemstoneRequest {
  seller_id: string;
  /** @minLength 1 */
  stone_type: string;
  /** @minimum 0 */
  carat: number;
  /** @minLength 1 */
  origin: string;
  /** @minLength 1 */
  treatment: string;
  /** GIA colour grade (D–M). Required when stone_type is diamond. */
  color?: string;
  /** GIA clarity grade. Required when stone_type is diamond. */
  clarity?: string;
  /** @minimum 0 */
  price: number;
  currency: Currency;
  /** @minLength 1 */
  certificate_number: string;
  /** @minLength 1 */
  image_url: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  buyer: "buyer",
  seller: "seller",
  admin: "admin",
} as const;

export type VerificationStatus =
  (typeof VerificationStatus)[keyof typeof VerificationStatus];

export const VerificationStatus = {
  unverified: "unverified",
  basic_verified: "basic_verified",
  premium_verified: "premium_verified",
} as const;

export type BoostPackType = (typeof BoostPackType)[keyof typeof BoostPackType];

export const BoostPackType = {
  boosts_5: "boosts_5",
  boosts_10: "boosts_10",
} as const;

export interface BoostListingRequest {
  seller_id: string;
}

export interface BuyBoostPackRequest {
  pack: BoostPackType;
}

export interface BoostResult {
  gemstone_id: string;
  is_featured: boolean;
  boost_expiry_date: string;
  boosts_remaining: number;
  boost_cost_usd: number;
  expires_in_days: number;
}

export interface BoostPackPurchaseResult {
  pack: BoostPackType;
  boosts_added: number;
  cost_usd: number;
  boosts_available: number;
  standard_price_per_boost: number;
  effective_price_per_boost: number;
}

export interface LaunchPricing {
  starter: number;
  growth: number;
  pro: number;
}

export interface LaunchVerificationFees {
  basic_verified: number;
  premium_verified: number;
}

export interface LaunchConfig {
  is_launch_period: boolean;
  verification_fees: LaunchVerificationFees;
  subscription_pricing: LaunchPricing;
  free_boosts_granted: number;
}

export interface SetLaunchPeriodRequest {
  is_launch_period: boolean;
}

export type SubscriptionPlan =
  (typeof SubscriptionPlan)[keyof typeof SubscriptionPlan];

export const SubscriptionPlan = {
  starter: "starter",
  growth: "growth",
  pro: "pro",
} as const;

export interface UpdateSubscriptionRequest {
  plan: SubscriptionPlan;
}

export type VerificationBadge =
  (typeof VerificationBadge)[keyof typeof VerificationBadge];

export const VerificationBadge = {
  none: "none",
  basic_verified: "basic_verified",
  premium_verified: "premium_verified",
} as const;

export type VerifiableTier =
  (typeof VerifiableTier)[keyof typeof VerifiableTier];

export const VerifiableTier = {
  basic_verified: "basic_verified",
  premium_verified: "premium_verified",
} as const;

export interface User {
  id: string;
  name: string;
  email: string;
  role: UserRole;
  verification_status: VerificationStatus;
  verification_fee_paid: boolean;
  verification_badge: VerificationBadge;
  requested_tier?: VerifiableTier;
  subscription_plan: SubscriptionPlan;
  /** @minimum 0 */
  free_boosts: number;
  is_founding_seller: boolean;
  created_at: string;
}

export interface VerificationRequestBody {
  tier: VerifiableTier;
}

export type AdminVerificationActionAction =
  (typeof AdminVerificationActionAction)[keyof typeof AdminVerificationActionAction];

export const AdminVerificationActionAction = {
  approve: "approve",
  reject: "reject",
} as const;

export interface AdminVerificationAction {
  action: AdminVerificationActionAction;
}

export interface CreateUserRequest {
  /** @minLength 1 */
  name: string;
  email: string;
  role: UserRole;
}

export interface ErrorResponse {
  error: string;
}

export type GetBuyerCreditSummaryParams = {
  /**
   * The seller's user ID making this request
   */
  requester_id: string;
};