# openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GemWorld B2B Gemstone Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transactions:
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a transaction request
      description: Buyer creates a new transaction for a gemstone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionRequest"
      responses:
        "201":
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /buyers/{id}/credit-summary:
    get:
      operationId: getBuyerCreditSummary
      tags: [transactions]
      summary: Get buyer credit summary (PRO sellers only)
      description: Returns credit score, total outstanding and overdue amounts for a buyer. Only accessible to PRO plan sellers.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: requester_id
          in: query
          required: true
          description: The seller's user ID making this request
          schema:
            type: string
      responses:
        "200":
          description: Full credit summary (PRO sellers only)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditSummary"
        "401":
          description: requester_id is missing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Requester is not a PRO seller
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Requester not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    TransactionStatus:
      type: string
      enum: [pending, completed, overdue]
    Transaction:
      type: object
      properties:
        id:
          type: string
        buyer_id:
          type: string
        seller_id:
          type: string
        inventory_id:
          type: string
        total_amount:
          type: number
          minimum: 0
        currency:
          $ref: "#/components/schemas/Currency"
        advance_paid:
          type: number
          minimum: 0
        credit_amount:
          type: number
          minimum: 0
        due_date:
          type: string
          format: date-time
        status:
          $ref: "#/components/schemas/TransactionStatus"
        created_at:
          type: string
          format: date-time
      required:
        - id
        - buyer_id
        - seller_id
        - inventory_id
        - total_amount
        - currency
        - advance_paid
        - credit_amount
        - due_date
        - status
        - created_at
    CreateTransactionRequest:
      type: object
      properties:
        buyer_id:
          type: string
        seller_id:
          type: string
        inventory_id:
          type: string
        total_amount:
          type: number
          minimum: 0
        currency:
          $ref: "#/components/schemas/Currency"
        advance_paid:
          type: number
          minimum: 0
        credit_amount:
          type: number
          minimum: 0
        due_date:
          type: string
          format: date-time
      required:
        - buyer_id
        - seller_id
        - inventory_id
        - total_amount
        - currency
        - advance_paid
        - credit_amount
        - due_date
...
    CreditSummary:
      type: object
      properties:
        buyer_id:
          type: string
        credit_score:
          type: integer
          minimum: 300
          maximum: 850
          description: Computed credit score (300–850). Higher is better.
        credit_rating:
          type: string
          description: Human-readable rating derived from credit score
        total_outstanding:
          type: number
        overdue_amount:
          type: number
        completed_transactions:
          type: integer
        overdue_transactions:
          type: integer
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
      required:
        - buyer_id
        - credit_score
        - credit_rating
        - total_outstanding
        - overdue_amount
        - completed_transactions
        - overdue_transactions
        - transactions
    Currency:
      type: string
      enum: [USD, INR, AED]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error